% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_triangles.R
\name{hmm_triangles}
\alias{hmm_triangles}
\title{Heightmap meshing}
\usage{
hmm_triangles(x, invert = FALSE, blur_sigma = 0, border_size = 0,
  border_height = 1, base_height = 0, max_error = 0.001,
  max_triangles = 0, max_points = 0, z_scale = 1,
  z_exaggeration = 1, quiet = FALSE, stl_file = "",
  normal_file = "")
}
\arguments{
\item{x}{matrix}

\item{invert}{logical, invert heightmap, default is \code{FALSE}}

\item{blur_sigma}{gaussian blur sigma, default is \code{0} for no blur}

\item{border_size}{border size in pixels, default is \code{0}}

\item{border_height}{border z height, default is \code{1}}

\item{base_height}{solid base height, default is no base \code{0}}

\item{max_error}{maximum triangulation error, default is \code{0.001}}

\item{max_triangles}{maximum number of triangles, default is no maximum \code{0}}

\item{max_points}{maximum number of vertices, default is no maximum \code{0}}

\item{z_scale}{z scale relative to x, y}

\item{z_exaggeration}{z exaggeration, default is \code{1}}

\item{quiet}{report metrics from triangulation, default is \code{TRUE}}

\item{stl_file}{if set the triangulation will be written to file, default is no write \code{""}}

\item{normal_file}{if set the normal map will be written to file, default is no write \code{""}}
}
\value{
nothing at the moment, an empty list
}
\description{
Create an efficient triangulation from a height map.
}
\details{
Currently input must be a matrix.
}
\examples{
f <- system.file("extdata/volcano1.png", package = "meshr", mustWork = TRUE)
im <- 255 * (volcano - min(volcano))/diff(range(volcano))
#tfile <- tempfile()
#meshr:::hmm_triangles(im, stl_file = tfile)
#tris <- rgl::readSTL(tfile, plot = FALSE)
}
