% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmr-package.R
\docType{package}
\name{meshroom-package}
\alias{meshroom}
\alias{meshroom-package}
\title{meshroom: Meshing Tools}
\description{
Meshes are awesome, we include some facilities for them. Heightmap meshing is an efficient triangulation of a raster height map using the 'hmm' library
 a modern implementation of a nice algorithm from the 1995 paper 'Fast Polygonal Approximation of Terrains and 
 Height Fields' by Garland and Heckbert <doi:10.1.1.70.9740>. The meshes produced by hmm satisfy the Delaunay condition 
 and can satisfy a specified maximal error or maximal number of triangles or vertices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fogleman/hmm}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com} (YOUR-ORCID-ID)

Authors:
\itemize{
  \item Michael Fogleman (author of hmm library)
}

}
\keyword{internal}
